added admin authorization for products

+++ app/Http/Controllers/ProductController.php
@@ -8,6 +8,7 @@ use App\Http\Resources\ProductResource;
 use App\Models\Product;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 
 class ProductController extends Controller
 {
@@ -21,11 +22,11 @@ class ProductController extends Controller
         return response([
 
             'data' => ProductCollection::collection(Product::paginate(5))
-   
-          ],Response::HTTP_PARTIAL_CONTENT);
+
+        ], Response::HTTP_PARTIAL_CONTENT);
     }
 
-  
+
     /**
      * Store a newly created resource in storage.
      *
@@ -34,22 +35,29 @@ class ProductController extends Controller
      */
     public function store(RequestsProduct $request)
     {
-        $product = new Product;
-        $product->name = $request->name;
-        $product->detail = $request->description;
-        $product->price = $request->price;
-        $product->stock = $request->stock;
-        $product->discount = $request->discount;
-        $product->admin_id =  request()->user()->id;
- 
-        $product->save();
- 
-        return response([
- 
-          'data' => new ProductResource($product)
- 
-        ],Response::HTTP_CREATED);
- 
+        if (request()->user()->isAdmin()) {
+            $product = new Product;
+            $product->name = $request->name;
+            $product->detail = $request->description;
+            $product->price = $request->price;
+            $product->stock = $request->stock;
+            $product->discount = $request->discount;
+            $product->admin_id =  request()->user()->id;
+
+            $product->save();
+
+            return response([
+
+                'data' => new ProductResource($product)
+
+            ], Response::HTTP_CREATED);
+        } else {
+            return response([
+
+                'message' => 'This user is unauthorized for this action'
+
+            ], Response::HTTP_UNAUTHORIZED);
+        }
     }
 
     /**
@@ -72,17 +80,25 @@ class ProductController extends Controller
      */
     public function update(Request $request, Product $product)
     {
-        $request['detail'] = $request->description;
+        if (request()->user()->isAdmin()) {
+            $request['detail'] = $request->description;
 
-        unset($request['description']);
+            unset($request['description']);
 
-        $product->update($request->all());
+            $product->update($request->all());
 
-       return response([
+            return response([
 
-         'data' => new ProductResource($product)
+                'data' => new ProductResource($product)
 
-       ],Response::HTTP_CREATED);
+            ], Response::HTTP_CREATED);
+        } else {
+            return response([
+
+                'message' => 'This user is unauthorized for this action'
+
+            ], Response::HTTP_UNAUTHORIZED);
+        }
     }
 
     /**
@@ -93,8 +109,16 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
-        $product->delete();
+        if (request()->user()->isAdmin()) {
+            $product->delete();
 
-        return response(null,Response::HTTP_NO_CONTENT);
+            return response(null, Response::HTTP_NO_CONTENT);
+        } else {
+            return response([
+
+                'message' => 'This user is unauthorized for this action'
+
+            ], Response::HTTP_UNAUTHORIZED);
+        }
     }
 }
